requirements.txt
fastapi
uvicorn[standard]
python-multipart
chromadb
python-dotenv
docling
docling-core
langchain-text-splitters
groq